fix: `composes` & relative path issue to deploy
in the `footer.module.css` module several `composes` properties were im-
porting from other modules with a relative path. turns out netlify does
not like that so the deployment failed. this commit temporarily imports
that other css module in the component and uses the styles from there.

this is only a workaround to not spend hours to try and fix that (and
as the composes were not important anyway) so perhaps in the future, a
proper fix might be pursued if needed.