fix: change how products were passed to `Product`
originally, there was a mistake in how the route was structured for pro-
duct page as an outlet, where the product page was listed as a child of
the shop page. this was changed and now the route expects a sibling ins-
tead. furthermore, the `ProductThumbNail` component now accepts the item
as a whole as its props, rather than destructured individual pieces of
it. this was done in order to pass the entire product object to the pro-
duct page to be displayed as a whole.