refactor: normalize the `Button` comp. PropTypes
the mentioned button component was an external StyledComponent imported
from an outsource repo. it was built using an older react version where
PropTypes were supported; however, this project uses react 19 which has
dropped support for PropTypes in favour of TypeScript which caused some
inconsistencies in this StyledComponent. this commit aims to normalize
the lack of PropTypes in a javascript environment, at least for the time
being.